change int to size_t

+++ src/proto.h
@@ -93,7 +93,7 @@ void dump_config(void);
 /*
 ** string.c functions
 */
-char *i18n_convstring(char *, char *, char *, int *);
+char *i18n_convstring(char *, char *, char *, size_t *);
 char *i18n_utf2numref(char *, int);
 char *i18n_numref2utf(char *);
